+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to SimpleBroker will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.1] - 2025-10-21
+### Added
+- Added --json as on option for the --status flag 
+
+
 ## [2.6.0] - 2025-10-21
 ### Added
 - Added --status flag that efficiently reports global information about the database